(GH-2) Prototype PsinderBlock Script
This commit adds a first prototype of a PsinderBlock script file as
a template as well as an example PsinderBlock script for managing the
TimeZone of a machine.

This implementation hews to the ideas in #2 around what a PsinderBlock
script might be. In particular, it ensures all custom logic for a
PsinderBlock script author lives in the *-PsinderBlock functions defined
in the begin block of the script with the exception of the resource's
properties, which must also be defined in the script's parameter block.